+++ app/src/components/index.js
@@ -1,11 +1,19 @@
 import Input from "./Input";
-import Button from "./Button";
+import Button from "./buttons/Button";
 import Spinner from "./Spinner";
-import FindButton from './FindButton'
+import FindButton from "./buttons/FindButton";
 import Image from "./Image";
 
 // mains
 import MarketContainer from "./mains/market/MarketContainer";
 import LinearContainer from "./mains/linear/LinearContainer";
 
-export { Input, Button, Spinner, FindButton, Image, MarketContainer, LinearContainer };
+export {
+  Input,
+  Button,
+  Spinner,
+  FindButton,
+  Image,
+  MarketContainer,
+  LinearContainer,
+};
